apps/face-recognition/requirements.txt
face_recognition_models
Click>=6.0
dlib>=19.3.0
numpy
flask
requests
Pillow
scipy>=0.17.0
numpy